Implement rechunking of streams to avoid data loss
