+++ sdk/src/runtime/ssrBridge.ts
@@ -8,7 +8,7 @@
 // `rwsdk/__ssr_bridge`. We have bundler logic (ssrBridgePlugin) that looks out
 // for imports to it.
 
-export { renderRscThenableToHtmlStream } from "./render/renderRscThenableToHtmlStream";
+export { renderHtmlStream } from "./render/renderHtmlStream";
 
 export {
   ssrLoadModule,
